refactor(pos) :  패키지 - product 패키지 변경
